Split agent readiness threshold helpers
